+++ app/src/main/java/org/sopt/soptseminar_week1/data/FollowingListInfo.kt
