+++ app/src/main/java/org/sopt/soptseminar_week1/view/HomeActivity.kt
@@ -1,10 +1,11 @@
-package org.sopt.soptseminar_week1
+package org.sopt.soptseminar_week1.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.result.contract.ActivityResultContracts
+import org.sopt.soptseminar_week1.data.RepositoryListInfo
 import org.sopt.soptseminar_week1.databinding.ActivityHomeBinding
 import org.sopt.soptseminar_week1.utils.activityLogger
