+++ app/src/main/java/org/sopt/soptseminar_week1/view/RepositoryListAdapter.kt
@@ -1,8 +1,9 @@
-package org.sopt.soptseminar_week1
+package org.sopt.soptseminar_week1.view
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import org.sopt.soptseminar_week1.data.RepositoryListInfo
 import org.sopt.soptseminar_week1.databinding.ItemRepositoryListBinding
 
 class RepositoryListAdapter :
